requirements.txt
pandas
selenium
beautifulsoup4
requests
python-dotenv
openai
tqdm
firebase-admin
faiss-cpu
pinecone-client
tiktoken